Fix wasm api usage -- need to use BigInt for our int64_t cleverness
Not sure if the int64_t cleverness is all that clever, as a result. May
reevaluate...

+++ src/lib/fountain/fountain_decoder_sink.h
@@ -138,7 +138,7 @@ public:
 		// -- and the assembled file will we dropped from RAM.
 		// but if no callback is provided, you can do something else.
 		store(md, s);
-		return md.id();
+		return (int64_t)md.id();
 	}
 
 	bool write(const char* data, unsigned length)
